Summarize the cereal and beverage totals
It is critical to this application not just to know how much has been
consumed since records began (arbitrarily 1000 records ago), we must
also be able to convert this into units that can be easily visualised by
the target audience.

A red brick is said to be between 2 and 3 kg, so we have taken the mean
value of 2.5kg.